pyproject.toml
[project]
name = "bytesapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.14",
    "ffmpeg>=1.4",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
